requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.4
django-livereload-server==0.4
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-texttospeech==2.11.0
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
gTTS==2.2.4
gunicorn==20.1.0
idna==3.3
proto-plus==1.20.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tornado==6.1
urllib3==1.26.9
